Vue 클릭 이벤트 설정
