+++ SideJobVue/src/components/Chicken_main.vue
@@ -17,7 +17,7 @@
 							{{i.content}}	
 						</div>
 						<div class="item-btn-wrap">
-							<button class="item-btn">Detail</button>
+							<button @click="viewCount(i.recr_id)" class="item-btn">Detail</button>
 						</div>
 					</div>
 				</div>
@@ -25,18 +25,29 @@
 		</div>
 	</div>
 </template>
+
 <script>
+import { eventBus } from '../main';
 
 export default {
   name: 'chickenMain',
-  data () {
+  data: function () {
     return {
-      items: [
-		  {contents: 'item'}
-	  ]
+      items: [],
+      img_src: 'http://www.urbanbrush.net/web/wp-content/uploads/edd/2018/02/web-20180211134119206041.png'
     }
-}
+  },
+  created () {
+    this.$axios.get('http://soorokim.duckdns.org:19223/getRecruitList')
+      .then((response) => {
+        this.items = response.data
+        eventBus.$emit('listChange', this.items.length);
+      })
+  },
+  methods: {
+		viewCount: function(id) {
+			console.log(id);
+    }
+	}
 }
 </script>
-<script src="../scripts/company_data.js"/>
-
